% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BioString.R
\name{matchID}
\alias{matchID}
\title{match two string vector}
\usage{
matchID(a, b)
}
\arguments{
\item{a}{The query string vector}

\item{b}{The target string vector}
}
\value{
the index in b for each query element in a
}
\description{
Search out the index in b for each element in a
}

