% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BioNetwork.R
\name{saveSif}
\alias{saveSif}
\title{save the adjacency matrix to sif network file,}
\usage{
saveSif(net, filepath, mode = "undirected", diag = FALSE, type = "pp")
}
\arguments{
\item{filepath}{full path of the networkfile}

\item{mode}{network type, mode=("undirected","directed")}

\item{diag}{if self loop is included}

\item{type}{edge type}

\item{network}{adjacency matrix}
}
\value{
no return
}
\description{
save the adjacency matrix using sif network format
}

