% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BioMath.R
\name{square}
\alias{square}
\title{Square a number}
\usage{
square(x)
}
\arguments{
\item{x}{A numeric value to be squred}
}
\value{
The square of the input
}
\description{
Take in any numeric value and squares it
}

